Fix warning

+++ Pod/Classes/YWeatherAPI.m
@@ -2400,7 +2400,7 @@ NSString* const kYWAPressureTrendRising = @"1";
 + (NSString *)urlencode:(NSString*) s {
     NSMutableString *output = [NSMutableString string];
     const unsigned char *source = (const unsigned char *)[s UTF8String];
-    int sourceLen = strlen((const char *)source);
+    unsigned long sourceLen = strlen((const char *)source);
     for (int i = 0; i < sourceLen; ++i) {
         const unsigned char thisChar = source[i];
         if (thisChar == ' '){
